+++ Changelog.md
@@ -594,6 +594,7 @@ Specific changes:
   * [tests] Add manual tests parsetests/.
   * [bin] Add: silly-opt -- an easy way to run mlir-opt against a silly dialect file.
   * Add lit test build infrastructure -- unfortunately depends on my llvm-build/ dir and llvm-project cmake configuration -- but it's a CI/CD start.
+  * [tests] ReturnOp verifier and some initial returnop dialect tests.  Move the {ScopeOp,DeclareOp}::verify out of line.
 
 ## tag: V7 (Jan 4, 2025)